Add tsdoc tags and typedocs generation
