+++ packages/cypher-builder/package.json
@@ -13,7 +13,8 @@
     ],
     "scripts": {
         "test": "jest",
-        "build": "tsc --build tsconfig.production.json"
+        "build": "tsc --build tsconfig.production.json",
+        "docs": "typedoc src/Cypher.ts"
     },
     "repository": {
         "type": "git",
@@ -37,6 +38,7 @@
         "expect-type": "0.15.0",
         "jest": "29.3.1",
         "ts-jest": "29.0.3",
+        "typedoc": "^0.23.21",
         "typescript": "4.9.3"
     }
 }
