+++ packages/cypher-builder/src/apoc/functions/RunFirstColumn.ts
@@ -23,6 +23,10 @@ import type { Variable } from "../../references/Variable";
 import type { CypherEnvironment } from "../../Environment";
 import type { MapExpr } from "../../expressions/map/MapExpr";
 
+/**
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export class RunFirstColumn extends CypherASTNode {
     private innerClause: Clause | string;
     private variables: Variable[] | MapExpr;
