+++ packages/cypher-builder/src/clauses/mixins/WithWhere.ts
@@ -22,12 +22,14 @@ import { Where } from "../sub-clauses/Where";
 import { and, BooleanOp } from "../../expressions/operations/boolean";
 import { PropertyRef } from "../../references/PropertyRef";
 import { ComparisonOp, eq } from "../../expressions/operations/comparison";
-import type { Predicate, VariableLike } from "../../types";
+import type { Predicate } from "../../types";
 import { Reference } from "../../references/Reference";
 import type { RelationshipRef } from "../../references/RelationshipRef";
 import type { NodeRef } from "../../references/NodeRef";
 import type { Variable } from "../../references/Variable";
+import type { Literal } from "../../references/Literal";
 
+export type VariableLike = Reference | Literal | PropertyRef;
 type VariableWithProperties = Variable | NodeRef | RelationshipRef | PropertyRef;
 
 export abstract class WithWhere extends ClauseMixin {
