+++ packages/cypher-builder/src/expressions/Case.ts
@@ -23,7 +23,10 @@ import { padBlock } from "../utils/pad-block";
 import { compileCypherIfExists } from "../utils/compile-cypher-if-exists";
 import type { Expr, Predicate } from "../types";
 
-/** Case statement <https://neo4j.com/docs/cypher-manual/current/syntax/expressions/#query-syntax-case> */
+/** Case statement
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/syntax/expressions/#query-syntax-case)
+ * @group Expressions
+ */
 export class Case<C extends Expr | undefined = undefined> extends CypherASTNode {
     private comparator: Expr | undefined;
     private whenClauses: When<C>[] = [];
