+++ packages/cypher-builder/src/expressions/Exists.ts
@@ -22,6 +22,10 @@ import type { Clause } from "../clauses/Clause";
 import { CypherASTNode } from "../CypherASTNode";
 import { padBlock } from "../utils/pad-block";
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/syntax/expressions/#existential-subqueries)
+ * @group Expressions
+ */
 export class Exists extends CypherASTNode {
     private subQuery: CypherASTNode;
