+++ packages/cypher-builder/src/expressions/HasLabel.ts
@@ -22,6 +22,14 @@ import type { CypherEnvironment } from "../Environment";
 import { escapeLabel } from "../utils/escape-label";
 import type { NodeRef } from "../references/NodeRef";
 
+/** Generates a predicate to check if a node has a label
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/syntax/expressions/#existential-subqueries)
+ * @group Expressions
+ * @example
+ * ```cypher
+ * MATCH(this) WHERE this:MyNode
+ * ```
+ */
 export class HasLabel extends CypherASTNode {
     private node: NodeRef;
     private expectedLabels: string[];
