+++ packages/cypher-builder/src/expressions/functions/CypherFunction.ts
@@ -22,10 +22,18 @@ import { CypherASTNode } from "../../CypherASTNode";
 import type { CypherEnvironment } from "../../Environment";
 import type { Expr } from "../../types";
 
+/** Represents a Cypher Function
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export class CypherFunction extends CypherASTNode {
     protected name: string;
     private params: Array<Expr>;
 
+    /**
+     * @hidden
+     */
     constructor(name: string, params: Array<Expr> = []) {
         super();
         this.name = name;
@@ -37,6 +45,9 @@ export class CypherFunction extends CypherASTNode {
         }
     }
 
+    /**
+     * @hidden
+     */
     public getCypher(env: CypherEnvironment): string {
         const argsStr = this.params.map((expr) => expr.getCypher(env)).join(", ");
 
@@ -44,68 +55,149 @@ export class CypherFunction extends CypherASTNode {
     }
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/scalar/#functions-coalesce)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function coalesce(expr: Expr, ...optionalExpr: Expr[]): CypherFunction {
     return new CypherFunction("coalesce", [expr, ...optionalExpr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/spatial/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function point(variable: Expr): CypherFunction {
     return new CypherFunction("point", [variable]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/4.3/functions/spatial/#functions-distance)
+ * @group Expressions
+ * @category Cypher Functions
+ * @deprecated No longer supported in Neo4j 5. Use {@link pointDistance} instead.
+ */
 export function distance(lexpr: Expr, rexpr: Expr): CypherFunction {
     return new CypherFunction("distance", [lexpr, rexpr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/spatial/#functions-distance)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function pointDistance(lexpr: Expr, rexpr: Expr): CypherFunction {
     return new CypherFunction("point.distance", [lexpr, rexpr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/list/#functions-labels)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function labels(nodeRef: Variable): CypherFunction {
     return new CypherFunction("labels", [nodeRef]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-datetime)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function cypherDatetime(): CypherFunction {
     return new CypherFunction("datetime");
 }
 
 // TODO: Add optional input to date functions - https://neo4j.com/docs/cypher-manual/current/functions/#header-query-functions-temporal-instant-types
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-date)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function cypherDate(): CypherFunction {
     return new CypherFunction("date");
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/scalar/#functions-coalesce)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function cypherLocalDatetime(): CypherFunction {
     return new CypherFunction("localdatetime");
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-localdatetime)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function cypherLocalTime(): CypherFunction {
     return new CypherFunction("localtime");
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/temporal/#functions-time)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function cypherTime(): CypherFunction {
     return new CypherFunction("time");
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-count)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function count(expr: Expr): CypherFunction {
     return new CypherFunction("count", [expr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-min)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function min(expr: Expr): CypherFunction {
     return new CypherFunction("min", [expr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-max)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function max(expr: Expr): CypherFunction {
     return new CypherFunction("max", [expr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-avg)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function avg(expr: Expr): CypherFunction {
     return new CypherFunction("avg", [expr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/aggregating/#functions-sum)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function sum(expr: Expr): CypherFunction {
     return new CypherFunction("sum", [expr]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/scalar/#functions-randomuuid)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function randomUUID(): CypherFunction {
     return new CypherFunction("randomUUID");
 }
