+++ packages/cypher-builder/src/expressions/functions/StringFunctions.ts
@@ -22,54 +22,119 @@ import { filterTruthy } from "../../utils/filter-truthy";
 
 import { CypherFunction } from "./CypherFunction";
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function left(original: Expr, length: Expr): CypherFunction {
     return new CypherFunction("left", [original, length]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function lTrim(original: Expr): CypherFunction {
     return new CypherFunction("lTrim", [original]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function replace(original: Expr, search: Expr, replace: Expr): CypherFunction {
     return new CypherFunction("replace", [original, search, replace]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function reverse(original: Expr): CypherFunction {
     return new CypherFunction("reverse", [original]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function right(original: Expr, length: Expr): CypherFunction {
     return new CypherFunction("right", [original, length]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function rTrim(original: Expr): CypherFunction {
     return new CypherFunction("rTrim", [original]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function split(original: Expr, delimiter: Expr): CypherFunction {
     return new CypherFunction("split", [original, delimiter]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function substring(original: Expr, start: Expr, length?: Expr): CypherFunction {
     return new CypherFunction("substring", filterTruthy([original, start, length]));
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function toLower(original: Expr): CypherFunction {
     return new CypherFunction("toLower", [original]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function toString(expression: Expr): CypherFunction {
     return new CypherFunction("toString", [expression]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function toStringOrNull(expression: Expr): CypherFunction {
     return new CypherFunction("toStringOrNull", [expression]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function toUpper(original: Expr): CypherFunction {
     return new CypherFunction("toUpper", [original]);
 }
 
+/**
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/functions/string/)
+ * @group Expressions
+ * @category Cypher Functions
+ */
 export function trim(original: Expr): CypherFunction {
     return new CypherFunction("trim", [original]);
 }
