+++ packages/cypher-builder/src/expressions/map/MapProjection.ts
@@ -22,7 +22,14 @@ import type { CypherCompilable, Expr } from "../../types";
 import type { Variable } from "../../references/Variable";
 import { serializeMap } from "../../utils/serialize-map";
 
-/** Represents a Map projection https://neo4j.com/docs/cypher-manual/current/syntax/maps/#cypher-map-projection */
+/** Represents a Map projection
+ * @see [Cypher Documentation](https://neo4j.com/docs/cypher-manual/current/syntax/maps/#cypher-map-projection)
+ * @group Expressions
+ * @example
+ * ```cypher
+ * this { .title }
+ * ```
+ */
 export class MapProjection implements CypherCompilable {
     private extraValues: Record<string, Expr>;
     private variable: Variable;
